+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-**Version 4.1.1** @ 13.04.2019
-- Updated the Scryfall database to: 2019-04-14 10:24 UTC.
+**Version 4.1.2** @ 13.04.2019
+- Updated the Scryfall database to: 2019-04-20 10:39 UTC.
+- Fixed long names and mana cost display in CardTableViewCell.
 
 **Version 4.0.9** @ 06.01.2019
 - Updated the Scryfall database to: 2019-02-06 10:27 UTC.